+++ core/app/models/comable/order_detail.rb
@@ -6,13 +6,17 @@ module Comable
     belongs_to :stock, class_name: Comable::Stock.name, foreign_key: Comable::Stock.table_name.singularize.foreign_key
     belongs_to :order_delivery, class_name: Comable::OrderDelivery.name, foreign_key: Comable::OrderDelivery.table_name.singularize.foreign_key
 
+    accepts_nested_attributes_for :stock
+
     # TODO: バリデーションの追加
 
     delegate :product, to: :stock
     delegate :guest_token, to: :order_delivery
     delegate :complete?, to: :order_delivery
+    delegate :order, to: :order_delivery
 
     before_save :save_to_add_cart, unless: :complete?
+    before_save :verify_quantity, unless: :complete?
 
     def save_to_complete
       self.attributes = current_attributes
@@ -39,10 +43,30 @@ module Comable
       price * quantity
     end
 
+    def valid_order_quantity?
+      if quantity <= 0
+        add_order_quantity_invalid_error_to_order
+        return false
+      end
+
+      if stock.soldout?(quantity: quantity)
+        add_product_soldout_error_to_order
+        return false
+      end
+
+      true
+    end
+
     private
 
     def decrement_stock
-      stock.decrement!(quantity: quantity)
+      return unless quantity
+      return unless stock.quantity
+      stock.quantity -= quantity
+    end
+
+    def verify_quantity
+      fail Comable::NoStock if stock.soldout?(quantity: quantity)
     end
 
     def current_attributes
@@ -55,6 +79,14 @@ module Comable
         sku_h_choice_name: stock.sku_h_choice_name,
         sku_v_choice_name: stock.sku_v_choice_name
       }
+    end
+
+    def add_order_quantity_invalid_error_to_order
+      order.errors.add :base, I18n.t('comable.errors.messages.order_quantity_invalid', name: stock.name_with_sku)
+    end
+
+    def add_product_soldout_error_to_order
+      order.errors.add :base, I18n.t('comable.errors.messages.product_soldout', name: stock.name_with_sku)
     end
   end
 end