+++ core/lib/comable/errors.rb
@@ -1,4 +1,7 @@
 module Comable
   class InvalidOrder < StandardError
   end
+
+  class NoStock < StandardError
+  end
 end