+++ frontend/app/controllers/comable/carts_controller.rb
@@ -1,30 +1,22 @@
 module Comable
   class CartsController < Comable::ApplicationController
-    def show
-    end
+    rescue_from Comable::NoStock, with: :no_stock
 
     def add
-      product = Comable::Product.where(id: params[:product_id]).first
-      return redirect_by_product_not_found unless product
+      cart_item = find_cart_item
+      return redirect_by_product_not_found unless cart_item
 
-      if product.sku?
-        stock = product.stocks.where(id: params[:stock_id]).first
-        return redirect_by_product_not_found unless stock
-      end
-
-      # TODO: 在庫確認
-      current_customer.add_cart_item(stock || product, quantity: params[:quantity].to_i)
+      current_customer.add_cart_item(cart_item, cart_item_options)
 
       flash[:notice] = I18n.t('comable.carts.add_product')
       redirect_to cart_path
     end
 
     def update
-      stock = Comable::Stock.where(id: params[:stock_id]).first
-      return redirect_by_product_not_found unless stock
+      cart_item = find_cart_item
+      return redirect_by_product_not_found unless cart_item
 
-      # TODO: 在庫確認
-      current_customer.reset_cart_item(stock, quantity: params[:quantity].to_i)
+      current_customer.reset_cart_item(cart_item, cart_item_options)
 
       flash[:notice] = I18n.t('comable.carts.update')
       redirect_to cart_path
@@ -33,8 +25,27 @@ module Comable
     private
 
     def redirect_by_product_not_found
-      flash[:error] = I18n.t('comable.carts.product_not_found')
+      flash[:alert] = I18n.t('comable.errors.messages.products_not_found')
       redirect_to :back
+    end
+
+    def no_stock
+      flash[:alert] = I18n.t('comable.errors.messages.products_soldout')
+      redirect_to cart_path
+    end
+
+    def find_cart_item
+      cart_item = Comable::Stock.where(id: params[:stock_id]).first
+      cart_item ||= Comable::Product.where(id: params[:product_id]).first
+      return unless cart_item
+      return if cart_item.is_a?(Comable::Product) && cart_item.sku?
+      cart_item
+    end
+
+    def cart_item_options
+      options = {}
+      options.update(quantity: params[:quantity].to_i) if params[:quantity]
+      options
     end
   end
 end