+++ frontend/app/controllers/comable/orders_controller.rb
@@ -7,11 +7,7 @@ module Comable
     before_filter :verify
     after_filter :save_order, except: :create
 
-    rescue_from ActiveRecord::RecordInvalid, with: :record_invalid
     rescue_from Comable::InvalidOrder, with: :order_invalid
-
-    def new
-    end
 
     def orderer
       case request.method_symbol
@@ -27,14 +23,11 @@ module Comable
       end
     end
 
-    def confirm
-    end
-
     def create
       order = current_customer.order
       if order.complete?
-        Comable::OrderMailer.complete(order).deliver if current_store.email_activate?
         flash[:notice] = I18n.t('comable.orders.success')
+        send_order_complete_mail
       else
         flash[:alert] = I18n.t('comable.orders.failure')
         redirect_to comable.confirm_order_path
@@ -42,6 +35,10 @@ module Comable
     end
 
     private
+
+    def send_order_complete_mail
+      Comable::OrderMailer.complete(@order).deliver if current_store.email_activate?
+    end
 
     def next_order_path(target_action_name = nil)
       case (target_action_name || action_name).to_sym
@@ -96,13 +93,8 @@ module Comable
       )
     end
 
-    def record_invalid
+    def order_invalid
       flash[:alert] = I18n.t('comable.orders.failure')
-      redirect_to comable.cart_path
-    end
-
-    def order_invalid(exception)
-      flash[:alert] = exception.message
       redirect_to comable.cart_path
     end
   end