+++ backend/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Comable 0.2.1 (October 29, 2014) ##
+
+*   No changes.
+
+
 ## Comable 0.2.0 (September 29, 2014) ##
 
 *   Implement feature for store settings.