+++ core/CHANGELOG.md
@@ -1,3 +1,11 @@
+*   Improve the order completion.
+
+
+## Comable 0.2.1 (October 29, 2014) ##
+
+*   Improve the error for no stock.
+
+
 ## Comable 0.2.0 (September 29, 2014) ##
 
 *   Add columns to order details.