+++ frontend/CHANGELOG.md
@@ -1,3 +1,13 @@
+*   Improve the order completion.
+
+
+## Comable 0.2.1 (October 29, 2014) ##
+
+*   Fix a bug that app side namespace is used.
+
+*   Improve the error for no stock.
+
+
 ## Comable 0.2.0 (September 29, 2014) ##
 
 *   Implement the order completion mail.